+++ src/org/moeaframework/analysis/plot/Plot.java
@@ -65,7 +65,6 @@ import org.jfree.data.xy.XYSeries;
 import org.jfree.data.xy.XYSeriesCollection;
 import org.moeaframework.Analyzer;
 import org.moeaframework.Analyzer.AnalyzerResults;
-import org.moeaframework.analysis.collector.Accumulator;
 import org.moeaframework.analysis.collector.Observation;
 import org.moeaframework.analysis.collector.Observations;
 import org.moeaframework.analysis.diagnostics.PaintHelper;
@@ -78,7 +77,7 @@ import org.moeaframework.core.Solution;
  * rapid creation of 2D plots, supporting:
  * <ol>
  *   <li>scatter plots of bi-objective populations,</li>
- *   <li>line plots of runtime dynamics (via an {@link Accumulator},</li>
+ *   <li>line plots of runtime dynamics (via an {@link Observations},</li>
  *   <li>box-and-whisker plots of performance statistics (via an {@link Analyzer}), and</li>
  *   <li>other plots of basic data types (e.g., line, scatter, area, stacked).</li>
  * </ol>
@@ -472,29 +471,6 @@ public class Plot {
 		
 		return this;
 	}
-
-	/**
-	 * Displays the runtime data stored in an {@link Accumulator} as one or
-	 * more line plots.
-	 * 
-	 * @param accumulator the {@code Accumulator} instance
-	 * @return a reference to this {@code Plot} instance
-	 * @deprecated use {@link #add(Observations)} instead
-	 */
-	@Deprecated
-	public Plot add(Accumulator accumulator) {
-		createXYPlot();
-		currentDataset++;
-		XYSeriesCollection dataset = new XYSeriesCollection();
-
-		for (String key : accumulator.keySet()) {
-			if (!key.equals("NFE")) {
-				add(key, accumulator, key, dataset);
-			}
-		}
-
-		return this;
-	}
 	
 	/**
 	 * Displays the runtime data stored in an {@link Observations} as one or
@@ -514,20 +490,6 @@ public class Plot {
 
 		return this;
 	}
-
-	/**
-	 * Displays the runtime data for the given metric as a line plot.
-	 * 
-	 * @param label the label for the series
-	 * @param accumulator the {@code Accumulator} instance
-	 * @param metric the name of the performance metric to plot
-	 * @return a reference to this {@code Plot} instance
-	 * @deprecated use {@link #add(String, Observations, String)} instead
-	 */
-	@Deprecated
-	public Plot add(String label, Accumulator accumulator, String metric) {
-		return add(label, accumulator, metric, null);
-	}
 	
 	/**
 	 * Displays the runtime data for the given metric as a line plot.
@@ -540,39 +502,6 @@ public class Plot {
 	public Plot add(String label, Observations observations, String metric) {
 		return add(label, observations, metric, null);
 	}
-
-	/**
-	 * Displays the runtime data for the given metric as a line plot.  The
-	 * series is added to the given dataset, or if {@code null} a new dataset
-	 * is created.
-	 * 
-	 * @param label the label for the series
-	 * @param accumulator the {@code Accumulator} instance
-	 * @param metric the name of the performance metric to plot
-	 * @param dataset the dataset, or {@code null} if a new dataset should be
-	 *        created
-	 * @return a reference to this {@code Plot} instance
-	 * @deprecated
-	 */
-	@Deprecated
-	private Plot add(String label, Accumulator accumulator, String metric, XYSeriesCollection dataset) {
-		List<Number> xs = new ArrayList<Number>();
-		List<Number> ys = new ArrayList<Number>();
-		
-		try {
-			for (int i = 0; i < accumulator.size("NFE"); i++) {
-				xs.add((Number)accumulator.get("NFE", i));
-				ys.add((Number)accumulator.get(metric, i));
-			}
-		} catch (ClassCastException e) {
-			return this;
-		}
-		
-		line(label, xs, ys, dataset);
-		setLabelsIfBlank("NFE", "Value");
-
-		return this;
-	}
 	
 	/**
 	 * Displays the runtime data for the given metric as a line plot.  The
@@ -580,7 +509,7 @@ public class Plot {
 	 * is created.
 	 * 
 	 * @param label the label for the series
-	 * @param accumulator the {@code Accumulator} instance
+	 * @param observations the {@code Observations} containing the data
 	 * @param metric the name of the performance metric to plot
 	 * @param dataset the dataset, or {@code null} if a new dataset should be
 	 *        created